Use datetime's TimeType and FixedTimespanSet types
datetime@e6012ba003a8f5949510dea0f9ff334234202b07 added the ability to have time zones created at runtime and used in the same way as any other time zone. This commit takes advantage of that by using the TimeType and FixedTimespanSet and FixedTimespan structures present in datetime, allowing compiled time zones to be used alongside the static zoneinfo ones.

The tzdump example was a casualty of this, as we no longer have the ability to look inside time zones to get at their transitions (this is coming, though)